requirements.txt
pytest==7.4.0
selene>=2.0.0rc9
black>=25.1.0
allure-pytest
allure-python-commons~=2.13.5
python-dotenv~=1.0.1
selenium~=4.29.0
faker